Extract data providers into a central class
This will come in handy for the next step, which is creating a UtilsTest
