+++ etc/qa/phpstan.neon
@@ -4,3 +4,8 @@ parameters:
 			classesAllowedToBeExtended:
 				- Psr\Log\AbstractLogger
 				- Psr\Log\Test\LoggerInterfaceTest
+	ignoreErrors:
+		-
+			identifier: shipmonk.deadMethod
+			path: ../../tests/DataProvider.php
+			count: 5
