+++ tests/CheckCorrectLogLevelTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace WyriHaximus\Tests\PSR3;
 
-use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\InvalidArgumentException;
@@ -12,20 +12,10 @@ use WyriHaximus\PSR3\Utils;
 
 use function WyriHaximus\PSR3\checkCorrectLogLevel;
 
-use const WyriHaximus\PSR3\LOG_LEVELS;
-
 final class CheckCorrectLogLevelTest extends TestCase
 {
-    /** @return iterable<array<string>> */
-    public static function provideCorrectLogLevels(): iterable
-    {
-        foreach (LOG_LEVELS as $logLevel) {
-            yield [$logLevel];
-        }
-    }
-
     #[Test]
-    #[DataProvider('provideCorrectLogLevels')]
+    #[DataProviderExternal(DataProvider::class, 'correctLogLevels')]
     public function correctLogLevel(string $logLevel): void
     {
         /** @phpstan-ignore function.deprecated */
@@ -33,24 +23,14 @@ final class CheckCorrectLogLevelTest extends TestCase
     }
 
     #[Test]
-    #[DataProvider('provideCorrectLogLevels')]
+    #[DataProviderExternal(DataProvider::class, 'correctLogLevels')]
     public function correctLogLevelUtils(string $logLevel): void
     {
         self::assertTrue(Utils::checkCorrectLogLevel($logLevel));
     }
 
-    /** @return iterable<array<string>> */
-    public static function provideInCorrectLogLevels(): iterable
-    {
-        yield ['yes'];
-
-        yield ['null'];
-
-        yield ['meltdown'];
-    }
-
     #[Test]
-    #[DataProvider('provideInCorrectLogLevels')]
+    #[DataProviderExternal(DataProvider::class, 'inCorrectLogLevels')]
     public function incorrectLogLevel(string $logLevel): void
     {
         self::expectException(InvalidArgumentException::class);
@@ -60,7 +40,7 @@ final class CheckCorrectLogLevelTest extends TestCase
     }
 
     #[Test]
-    #[DataProvider('provideInCorrectLogLevels')]
+    #[DataProviderExternal(DataProvider::class, 'inCorrectLogLevels')]
     public function incorrectLogLevelUtils(string $logLevel): void
     {
         self::expectException(InvalidArgumentException::class);
