+++ tests/ProcessPlaceHoldersTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace WyriHaximus\Tests\PSR3;
 
-use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use WyriHaximus\PSR3\Utils;
@@ -13,31 +13,9 @@ use function WyriHaximus\PSR3\processPlaceHolders;
 
 final class ProcessPlaceHoldersTest extends TestCase
 {
-    /** @return iterable<array<int, array<string, string>|string>> */
-    public static function provideForTestProcessPlaceHolders(): iterable
-    {
-        yield [
-            'foo.bar',
-            [],
-            'foo.bar',
-        ];
-
-        yield [
-            'foo.{var}',
-            ['var' => 'bar'],
-            'foo.bar',
-        ];
-
-        yield [
-            'foo.{var}',
-            ['voor' => 'bar'],
-            'foo.{var}',
-        ];
-    }
-
     /** @param array<string, mixed> $context */
     #[Test]
-    #[DataProvider('provideForTestProcessPlaceHolders')]
+    #[DataProviderExternal(DataProvider::class, 'processPlaceHolders')]
     public function processPlaceHolders(string $message, array $context, string $expectedOutput): void
     {
         /** @phpstan-ignore function.deprecated */
@@ -46,7 +24,7 @@ final class ProcessPlaceHoldersTest extends TestCase
 
     /** @param array<string, mixed> $context */
     #[Test]
-    #[DataProvider('provideForTestProcessPlaceHolders')]
+    #[DataProviderExternal(DataProvider::class, 'processPlaceHolders')]
     public function processPlaceHoldersUtils(string $message, array $context, string $expectedOutput): void
     {
         self::assertSame($expectedOutput, Utils::processPlaceHolders($message, $context));
